A bunch of changes from code review (#8)
* Butterworth/polynomialCoefficients: remove unnecessary copy

coeffs is not needed after the copy, but the compiler cannot optimize
it away because it's used afterwards

* Butterworth/polynomialCoefficients: remove unneeded sort

The elements are already sorted a few lines above, and remove_if
preserve the relative ordering.

* Butterworth/polynomialCoefficients: remove suspiciously looking .assign()

Assigning to a container using iterators from the same container
is not guaranteed to work, because assign() could be just clear()
+ push_back() and the iterators would be invalidated half-way through.

It seems the goal of the assign() is to trim the vector to remove
elements that were shifted by remove_if. We can obtain that result
with .resize(), which is also O(1) instead of O(n)

* Butterworth: fix suboptimal code

Avoid excessive copies and quadratic behavior

* Butterworth: optimize binomial coefficient computation

By merging it into the loop that calls choose() multiple times and
computing them incrementally.

* Butterworth: use integer math instead of double math

This is assuming everything will fit in 32 bit, which is not a given
for binomial coefficients...

* Butterworth: more integer math and double math instead of complex math

* Butterworth: reserve storage space upfront for vectors

Avoids quadratic behavior

* Allocate the riesz pyramids inline

There is no need to heap allocate them, the std::vector storage will
be heap-allocated if needed anyway.