Use a long running worker threads instead of std::async
std::async spawns a new thread for every call, which is wasteful.
Instead, use a fixed size thread pool and push new items using a queue.

For practicality, the queue is just std::queue + std::mutex, instead
of something fancy like a lock-free queue or RCU. It is unlikely
there will be contention so it does not matter.