Remove ComplexMat completely
In the process of doing this, I also changed itsR to a compact
representation of real/imaginary, but that causes a copy when
filtering in .build(), which is not really a win.

The copy could be avoided by handcoding the filter (which would be
a huge win, because the compiler can unroll the kernel loop
and optimize away the multiplication by 0), but the handcoded filter
code is not working as it should.