VideoSource: switch to native YUV/YCbCr
All our operations are in the Y channel (grayscale), so we don't
even need to read Cb/Cr from device memory, let alone spend precious
cycles converting to rgb (in hw), then to ycbcr (in sw),
then to rgb again (in sw) then to y again (in sw).

The tricky part is that YUV is a planar format, not a packed format.
That means that you get the Y component of all pixels, followed by
the Cb of all pixels, followed by the Cr of all pixels (as opposed
to R-G-B for each pixel). This is good for us, but also makes the
V4L2 API a little more annoying.